+++ weaver/datatype.py
@@ -75,6 +75,7 @@ from weaver.utils import (
     get_log_fmt,
     get_path_kvp,
     get_settings,
+    is_valid_url,
     now,
     request_extra
 )
@@ -1079,10 +1080,10 @@ class Job(Base, LoggerHandler):
 
     @accept_type.setter
     def accept_type(self, content_type):
-        # type: (Optional[Union[str]]) -> None
-        if not isinstance(content_type, str):
+        # type: (Optional[str]) -> None
+        if not (isinstance(content_type, str) or content_type is None):
             raise TypeError(f"Type 'str' is required for '{self.__name__}.accept_type'")
-        self["accept_type"] = content_type
+        self["accept_type"] = content_type or None
 
     @property
     def accept_language(self):
@@ -1091,10 +1092,24 @@ class Job(Base, LoggerHandler):
 
     @accept_language.setter
     def accept_language(self, language):
-        # type: (Optional[Union[str]]) -> None
-        if not isinstance(language, str):
+        # type: (Optional[str]) -> None
+        if not (isinstance(language, str) or language is None):
             raise TypeError(f"Type 'str' is required for '{self.__name__}.accept_language'")
-        self["accept_language"] = language
+        self["accept_language"] = language or None
+
+    @property
+    def accept_profile(self):
+        # type: () -> Optional[str]
+        return self.get("accept_profile")
+
+    @accept_profile.setter
+    def accept_profile(self, profile):
+        # type: (Optional[str]) -> None
+        if not (isinstance(profile, str) or profile is None):
+            raise TypeError(f"Type 'str' is required for '{self.__name__}.accept_profile'")
+        if profile and not is_valid_url(profile):
+            raise ValueError(f"Valid URI is required for '{self.__name__}.accept_profile'")
+        self["accept_profile"] = profile or None
 
     @property
     def execute_async(self):
@@ -1590,6 +1605,7 @@ class Job(Base, LoggerHandler):
         if self_link in ["status", None]:
             job_links.extend([
                 {"href": job_list, "rel": "collection", "title": "List of submitted jobs."},  # IANA
+                {"href": sd.OGC_API_PROC_PROFILE_JOB_DESC_URL, "rel": "profile", "title": "Job response profile."},
             ])
 
         if self.status in JOB_STATUS_CATEGORIES[StatusCategory.FINISHED]:
@@ -1612,10 +1628,17 @@ class Job(Base, LoggerHandler):
                     "href": f"{job_url}/exceptions", "rel": "http://www.opengis.net/def/rel/ogc/1.0/exceptions",
                     "title": "List of job exceptions if applicable in case of failing job."
                 })
-        job_links.append({
-            "href": f"{job_url}/logs", "rel": "logs",  # unofficial
-            "title": "List of collected job logs during process execution."
-        })
+        job_links.extend([
+            {
+                "href": f"{job_url}/logs", "rel": "logs",  # unofficial
+                "title": "List of collected job logs during process execution."
+            },
+            {
+                # official, same as 'rel="[ogc-rel:log]"'
+                "href": f"{job_url}/logs", "rel": "http://www.opengis.net/def/rel/ogc/1.0/log",
+                "title": "List of collected job logs during process execution."
+            }
+        ])
         if self_link in ["status", "inputs", "outputs", "results", "logs", "exceptions", "provenance"]:
             self_link_body = list(filter(lambda _link: _link["rel"].endswith(self_link), job_links))[-1]
             self_link_body = copy.deepcopy(self_link_body)
@@ -1709,6 +1732,7 @@ class Job(Base, LoggerHandler):
             "subscribers": self.subscribers,
             "accept_type": self.accept_type,
             "accept_language": self.accept_language,
+            "accept_profile": self.accept_profile,
         }
 
 
@@ -2718,6 +2742,7 @@ class Process(Base):
         proc_self = f"{proc_list}/{self.tag}" if self.version else proc_desc
         links = [
             {"href": proc_self, "rel": "self", "title": "Current process description."},
+            {"href": sd.OGC_API_PROC_PROFILE_PROC_DESC_URL, "rel": "profile", "title": "Process response profile."},
             {"href": f"{proc_desc}?f=xml", "rel": "alternate",
              "title": "Alternate process description.", "type": ContentType.APP_XML},
         ]