+++ weaver/wps_restapi/jobs/utils.py
@@ -66,13 +66,13 @@ from weaver.utils import (
     get_href_headers,
     get_path_kvp,
     get_request_args,
+    get_response_profile,
     get_sane_name,
     get_secure_path,
     get_settings,
     get_weaver_url,
     is_uuid,
-    make_link_header,
-    parse_kvp
+    make_link_header
 )
 from weaver.visibility import Visibility
 from weaver.wps.utils import (
@@ -108,6 +108,8 @@ if TYPE_CHECKING:
         HTTPValid,
         JobValueFormat,
         JSON,
+        Link,
+        ProcessExecution,
         PyramidRequest,
         SettingsType
     )
@@ -211,8 +213,8 @@ def get_job(request):
     return job
 
 
-def get_job_list_links(job_total, filters, request):
-    # type: (int, Dict[str, AnyValueType], AnyRequestType) -> List[JSON]
+def get_job_list_links(job_total, filters, grouped, request):
+    # type: (int, Dict[str, AnyValueType], Any, AnyRequestType) -> List[Link]
     """
     Obtains a list of all relevant links for the corresponding job listing defined by query parameter filters.
 
@@ -259,14 +261,20 @@ def get_job_list_links(job_total, filters, request):
     if cur_page < 0 or cur_page > max_page:
         raise IndexError(f"Page index {cur_page} is out of range from [0,{max_page}].")
 
-    alt_links = []
+    links = []
     if alt_path:
-        alt_links = [{
+        links.append({
             "href": get_path_kvp(alt_path, page=cur_page, **alt_kvp), "rel": "alternate",
             "type": ContentType.APP_JSON, "title": "Alternate endpoint with equivalent set of filtered jobs."
-        }]
+        })
 
-    links = alt_links + [
+    if not grouped:
+        links.append({
+            "href": sd.OGC_API_PROC_PROFILE_JOB_LIST_URL, "rel": "profile",
+            "title": "OGC API - Processes - Job List Profile reference."
+        })
+
+    links.extend([
         {"href": job_path, "rel": "collection",
          "type": ContentType.APP_JSON, "title": "Complete job listing (no filtering queries applied)."},
         {"href": base_url + sd.jobs_service.path, "rel": "search",
@@ -281,7 +289,7 @@ def get_job_list_links(job_total, filters, request):
          "type": ContentType.APP_JSON, "title": "First page of job query listing."},
         {"href": get_path_kvp(job_path, page=max_page, **kvp_params), "rel": "last",
          "type": ContentType.APP_JSON, "title": "Last page of job query listing."},
-    ]
+    ])
     if cur_page > 0:
         links.append({
             "href": get_path_kvp(job_path, page=cur_page - 1, **kvp_params), "rel": "prev",
@@ -342,19 +350,18 @@ def get_job_status_schema(request):
     :raises HTTPBadRequest: If an invalid combination :term:`Profile` and :term:`Media-Type` is requested.
     """
 
-    def make_headers(resolved_schema, content_type):
-        # type: (JobStatusProfileSchemaType, AnyContentType) -> HeadersType
+    def make_headers(resolved_schema, resolved_profile, content_type):
+        # type: (JobStatusProfileSchemaType, Optional[str], AnyContentType) -> HeadersType
         if content_type == ContentType.TEXT_HTML:
             return {"Content-Type": content_type}
         if content_type == ContentType.ANY and resolved_schema != JobStatusProfileSchema.WPS:
             content_type = ContentType.APP_JSON
-        content_profile = f"{content_type}; profile={resolved_schema}"
+        if resolved_profile == sd.OGC_API_PROC_PROFILE_JOB_DESC_URL:
+            content_profile = f"{content_type}; profile=\"{resolved_profile}\""
+        else:
+            content_profile = f"{content_type}; profile={resolved_schema}"
         content_headers = {"Content-Type": content_profile}
-        if resolved_schema == JobStatusProfileSchema.OGC:
-            content_headers["Content-Schema"] = sd.OGC_API_SCHEMA_JOB_STATUS_URL
-        elif resolved_schema == JobStatusProfileSchema.OPENEO:
-            content_headers["Content-Schema"] = sd.OPENEO_API_SCHEMA_JOB_STATUS_URL
-        elif content_type in ContentType.ANY_XML:
+        if content_type in ContentType.ANY_XML:
             if resolved_schema not in [JobStatusProfileSchema.WPS, None]:
                 raise HTTPBadRequest(
                     json={
@@ -365,29 +372,42 @@ def get_job_status_schema(request):
                 )
             content_headers["Content-Type"] = f"{ContentType.APP_XML}; profile={JobStatusProfileSchema.WPS}"
             content_headers["Content-Schema"] = sd.OGC_WPS_1_SCHEMA_JOB_STATUS_URL
+        else:
+            if resolved_schema == JobStatusProfileSchema.OGC:
+                content_headers["Content-Schema"] = sd.OGC_API_SCHEMA_JOB_STATUS_URL
+                content_headers["Content-Profile"] = sd.OGC_API_PROC_PROFILE_JOB_DESC_URL
+                content_headers["Link"] = make_link_header(sd.OGC_API_PROC_PROFILE_JOB_DESC_URL, rel="profile")
+            elif resolved_schema == JobStatusProfileSchema.OPENEO:
+                content_headers["Content-Schema"] = sd.OPENEO_API_SCHEMA_JOB_STATUS_URL
         return content_headers
+
+    profile = get_response_profile(request)
+    if profile == sd.OGC_API_PROC_PROFILE_JOB_DESC_URL:
+        schema = sd.JobStatusProfileSchema.OGC
+    elif profile == sd.OGC_WPS_1_SCHEMA_JOB_STATUS_URL:
+        schema = sd.JobStatusProfileSchema.WPS
+    elif profile == sd.OPENEO_API_SCHEMA_JOB_STATUS_URL:
+        schema = sd.JobStatusProfileSchema.OPENEO
+    else:
+        params = get_request_args(request)
+        schema = JobStatusProfileSchema.get(profile or params.get("schema"))
 
     content_accept = get_header("Accept", request.headers) or ContentType.APP_JSON
     content_media_type = clean_media_type_format(content_accept.split(",")[0], strip_parameters=True)
-
-    params = get_request_args(request)
-    schema = JobStatusProfileSchema.get(params.get("profile") or params.get("schema"))
     if schema:
-        headers = make_headers(schema, content_media_type)
+        headers = make_headers(schema, profile, content_media_type)
         return schema, headers
 
-    params = parse_kvp(content_accept)
-    profile = params.get("profile")
     if not profile:
         if content_media_type in ContentType.ANY_XML:
             schema = JobStatusProfileSchema.WPS
         else:
             schema = JobStatusProfileSchema.OGC
-        headers = make_headers(schema, content_media_type)
+        headers = make_headers(schema, profile, content_media_type)
         return schema, headers
 
-    schema = JobStatusProfileSchema.get(profile[0], default=JobStatusProfileSchema.OGC)
-    headers = make_headers(schema, content_media_type)
+    schema = JobStatusProfileSchema.get(profile, default=JobStatusProfileSchema.OGC)
+    headers = make_headers(schema, profile, content_media_type)
     return schema, headers
 
 
@@ -491,14 +511,14 @@ def get_results(  # pylint: disable=R1260
     wps_url = get_wps_output_url(settings)
     if not wps_url.endswith("/"):
         wps_url = f"{wps_url}/"
-    schema = JobInputsOutputsSchema.get(str(schema).lower(), default=JobInputsOutputsSchema.OLD)
+    schema = str(JobInputsOutputsSchema.get(str(schema).lower(), default=JobInputsOutputsSchema.OLD))
     strict = schema.endswith("+strict")
-    schema = schema.split("+")[0]
+    schema = schema.split("+", 1)[0]
     ogc_api = schema == JobInputsOutputsSchema.OGC
     outputs = {} if ogc_api else []
     fmt_key = "mediaType" if ogc_api else "mimeType"
     references = {}
-    for result in job.results:
+    for result in job.results:  # type: Union[JSON, JobValueItem]
         # Filter outputs not requested, unless 'all' requested by omitting
         out_id = get_any_id(result)
         if (
@@ -601,7 +621,7 @@ def get_results(  # pylint: disable=R1260
 
 def get_job_return(
     job=None,       # type: Optional[Job]
-    body=None,      # type: Optional[JSON]
+    body=None,      # type: Optional[ProcessExecution]
     headers=None,   # type: Optional[AnyHeadersContainer]
 ):                  # type: (...) -> Tuple[AnyExecuteResponse, AnyExecuteReturnPreference]
     """
@@ -657,6 +677,7 @@ def get_job_results_response(
     job,                        # type: Job
     *,                          # force named keyword arguments after
     container,                  # type: AnySettingsContainer
+    request=None,               # type: Optional[AnyRequestType]
     request_headers=None,       # type: Optional[AnyHeadersContainer]
     response_headers=None,      # type: Optional[AnyHeadersContainer]
     results_headers=None,       # type: Optional[AnyHeadersContainer]
@@ -694,7 +715,11 @@ def get_job_results_response(
         - :ref:`proc_exec_results`
 
     :param job: Job for which to generate the results response, which contains the originally submitted parameters.
-    :param container: Application settings.
+    :param container:
+        Request that was submitted for this response generation or an application settings container.
+        If the object cannot be the request directly, additional parameters must be provided to resolve relevant
+        properties defined in the request, such as control headers that can affect the response format.
+    :param request: Request object to look for more parameters that could affect the generation of this response.
     :param request_headers: Original headers submitted to the request that leads to this response.
     :param response_headers: Additional headers to provide in the response.
     :param results_headers: Headers that override originally submitted job parameters when requesting results.
@@ -703,9 +728,12 @@ def get_job_results_response(
     raise_job_dismissed(job, container)
     raise_job_bad_status_success(job, container)
     settings = get_settings(container)
+    request = request or container
+    request = None if isinstance(request, dict) else request
 
     results, _ = get_results(
-        job, container,
+        job,
+        container,
         value_key="value",
         schema=JobInputsOutputsSchema.OGC,  # not strict to provide more format details
         # no link headers since they are represented differently based on request parameters
@@ -737,9 +765,11 @@ def get_job_results_response(
     )
 
     headers = update_preference_applied_return_header(job, request_headers, headers)
+    profile = get_response_profile(request, request_headers)
+    is_doc_results = profile == sd.OGC_API_PROC_PROFILE_RESULTS_URL
 
-    # document/minimal response
-    if not is_raw and not is_accept_multipart and not is_single_output_minimal:
+    # document/minimal response, unless explicitly requested by profile content negotiation
+    if is_doc_results or (not is_raw and not is_accept_multipart and not is_single_output_minimal):
         try:
             results_schema = sd.ResultsDocument()
             results_json = results_schema.deserialize(results)
@@ -767,6 +797,15 @@ def get_job_results_response(
         # use deserialized contents such that only the applicable fields remain
         # (simplify compares, this is assumed by the following call)
         results_json = get_job_results_document(job, results_json, settings=settings)
+        headers.extend([
+            ("Content-Profile", sd.OGC_API_PROC_PROFILE_RESULTS_URL),
+            ("Link", make_link_header(sd.OGC_API_PROC_PROFILE_RESULTS_URL, rel="profile")),
+        ])
+        if is_doc_results:
+            # media-type is extended only if explicitly requested to avoid breaking clients relying on plain JSON
+            headers.update([
+                ("Content-Type", f"{ContentType.APP_JSON}; profile=\"{sd.OGC_API_PROC_PROFILE_RESULTS_URL}\"")
+            ])
         return HTTPOk(json=results_json, headers=headers)
 
     if not results:  # avoid schema validation error if all by reference
@@ -774,6 +813,7 @@ def get_job_results_response(
         # see:
         #   - https://docs.ogc.org/is/18-062r2/18-062r2.html#req_core_process-execute-sync-raw-ref
         #   - https://docs.ogc.org/DRAFTS/18-062.html#req_core_job-results-param-outputs-empty
+        #   - /req/core/process-execute-success-sync-empty-outputs
         return HTTPNoContent(headers=headers)
 
     # raw response can be data-only value, link-only or a mix of them
@@ -1390,9 +1430,10 @@ def raise_job_bad_status_success(job, container=None):
                 headers=headers,
                 json={
                     "title": "JobResultsFailed",
-                    "type": err_code,
+                    "type": "http://www.opengis.net/def/exceptions/ogcapi-processes-1/1.0/result-not-available",
                     "detail": "Job results not available because execution failed.",
                     "status": HTTPBadRequest.code,
+                    "error": err_code,
                     "cause": err_info,
                     "links": links
                 }
@@ -1429,7 +1470,7 @@ def raise_job_dismissed(job, container=None):
             headers=headers,
             json={
                 "title": "JobDismissed",
-                "type": "JobDismissed",
+                "type": "http://www.opengis.net/def/exceptions/ogcapi-processes-1/1.0/result-not-available",
                 "status": JobGone.code,
                 "detail": "Job was dismissed and artifacts have been removed.",
                 "cause": {"status": job.status},